+++ G26NewProject/MorosovaProj/src/test/java/parentTest/ParentTest.java
@@ -1,0 +1,39 @@
+package parentTest;
+
+import org.junit.After;
+import org.junit.Before;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+
+
+import java.io.File;
+import java.util.concurrent.TimeUnit;
+
+
+/**
+ * Created by zuzu on 7/17/2017.
+ */
+public class ParentTest
+{
+    WebDriver webDriver;
+
+    public ParentTest(WebDriver webDriver) {
+        this.webDriver = webDriver;
+    }
+
+    @Before
+    public void setUp()
+    {
+        File fileFF=new File(".././drivers/chromedriver.exe");
+        System.setProperty("webdriver.chrome.driver",fileFF.getAbsolutePath());
+        webDriver=new ChromeDriver();
+        webDriver.manage().window().maximize();
+        webDriver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
+    }
+
+    @After
+    public void tearDown()
+    {
+        webDriver.quit();
+    }
+}
